Renames folders and adds folders in preparation to the refactor to MVVM. Also refactors the settings view into smaller components.
