+++ Bean Juice/Models/Method.swift
