+++ Bean Juice/Views/BJTabView.swift
@@ -1,6 +1,6 @@
 //
-//  ContentView.swift
-//  Morning Brew
+//  BJTabView.swift
+//  Bean Juice
 //
 //  Created by Iiro Alhonen on 10.3.2020.
 //  Copyright © 2020 Nifty Tree Studios. All rights reserved.
@@ -8,19 +8,17 @@
 
 import SwiftUI
 
-/// The main content view. Contains the ``TabView``.
-struct ContentView: View {
+/// Contains the ``TabView``.
+struct BJTabView: View {
     @State private var selection = "Methods"
-    @AppStorage(wrappedValue: 150, "cupSize") var cupSize: Double
     @AppStorage(wrappedValue: false, "mlSelected") var mlSelected: Bool
-    @State private var ratio: Double = 13
-    @State private var cups: Double = 1
+    @AppStorage(wrappedValue: 150, "cupSize") var cupSize: Double
 
     @State private var customColor: Color = Color.blue
 
     var body: some View {
         TabView(selection: $selection) {
-            MethodSelectionView(cupSize: $cupSize, ratio: self.$ratio, cups: self.$cups, customColor: self.$customColor, mlSelected: $mlSelected)
+            MethodSelectionView(cupSize: $cupSize, customColor: self.$customColor, mlSelected: $mlSelected)
                 .tabItem {
                     VStack {
                         Image(systemName: "ellipsis.circle")
@@ -37,7 +35,7 @@ struct ContentView: View {
                     }
                 }
 
-            SettingsView(mlSelected: $mlSelected, cupSize: $cupSize, customColor: $customColor)
+            SettingsView(mlSelected: $mlSelected, customColor: $customColor)
                 .tabItem {
                     VStack {
                         Image(systemName: "gear")
