+++ Bean Juice/Views/Components/ItemRow.swift
@@ -12,8 +12,6 @@ struct ItemRowView: View {
     let method: Method
 
     @Binding var cupSize: Double
-    @Binding var ratio: Double
-    @Binding var cups: Double
     @Binding var customColor: Color
     @Binding var mlSelected: Bool
 
@@ -22,8 +20,6 @@ struct ItemRowView: View {
         NavigationLink(
             destination: MethodView(
                 cupSize: self.$cupSize,
-                ratio: self.$ratio,
-                cups: self.$cups,
                 method: method,
                 mlSelected: $mlSelected,
                 customColor: self.$customColor
